feature(charConditionals): added topaz
