+++ src/lib/characterConditionals.js
@@ -18,6 +18,7 @@ import yanqing from 'lib/conditionals/character/Yanqing';
 import welt from 'lib/conditionals/character/Welt';
 import firetrailblazer from 'lib/conditionals/character/TrailblazerPreservation';
 import physicaltrailblazer from 'lib/conditionals/character/TrailblazerDestruction';
+import topaz from 'lib/conditionals/character/Topaz';
 
 const Stats = Constants.Stats
 
@@ -72,148 +73,6 @@ export const characterOptionMapping = {
 }
 
 
-// function physicaltrailblazer(e) {
-//   const talentAtkScalingValue = talent(e, 0.20, 0.22)
-
-//   const basicScaling = basic(e, 1.00, 1.10)
-//   const skillScaling = skill(e, 1.25, 1.375)
-//   const ultScaling = ult(e, 4.5, 4.80)
-//   const ultEnhancedScaling = ult(e, 2.70, 2.88)
-
-//   return {
-//     display: () => (
-//       <Flex vertical gap={10} >
-//         <FormSwitch name='enhancedUlt' text='Aoe ult' />
-//         <FormSlider name='talentStacks' text='Talent stacks' min={0} max={2} />
-//       </Flex>
-//     ),
-//     defaults: () => ({
-//       enhancedUlt: true,
-//       talentStacks: 2,
-//     }),
-//     precomputeEffects: (request) => {
-//       const r = request.characterConditionals
-//       const x = Object.assign({}, baseComputedStatsObject)
-
-//       // Stats
-//       x[Stats.ATK_P] += r.talentStacks * talentAtkScalingValue
-//       x[Stats.DEF_P] += r.talentStacks * 0.10
-//       x[Stats.CR] += (request.enemyWeaknessBroken) ? 0.25 : 0
-
-//       // Scaling
-//       x.BASIC_SCALING += basicScaling
-//       x.SKILL_SCALING += skillScaling
-//       x.ULT_SCALING += (r.enhancedUlt) ? ultEnhancedScaling : ultScaling
-
-//       // Boost
-//       x.SKILL_BOOST += 0.25
-//       x.ULT_BOOST += (r.enhancedUlt) ? 0.25 : 0
-
-//       return x
-//     },
-//     calculateBaseMultis: (c) => {
-//       const x = c.x
-
-//       x.BASIC_DMG += x.BASIC_SCALING * x[Stats.ATK]
-//       x.SKILL_DMG += x.SKILL_SCALING * x[Stats.ATK]
-//       x.ULT_DMG += x.ULT_SCALING * x[Stats.ATK]
-//     }
-//   }
-// }
-
-function topaz(e) {
-  const proofOfDebtFuaVulnerability = skill(e, 0.50, 0.55)
-  const enhancedStateFuaScalingBoost = ult(e, 1.50, 1.65)
-  const enhancedStateFuaCdBoost = ult(e, 0.25, 0.275)
-
-  const basicScaling = basic(e, 1.00, 1.10)
-  const skillScaling = skill(e, 1.50, 1.65)
-  const fuaScaling = talent(e, 1.50, 1.65)
-
-  // 0.06
-  const basicHitCountMulti = ASHBLAZING_ATK_STACK *
-    (1 * 1 / 1)
-
-  // 0.18
-  const fuaHitCountMulti = ASHBLAZING_ATK_STACK *
-    (1 * 1 / 7 + 2 * 1 / 7 + 3 * 1 / 7 + 4 * 1 / 7 + 5 * 1 / 7 + 6 * 1 / 7 + 7 * 1 / 7)
-
-  // 0.252
-  const fuaEnhancedHitCountMulti = ASHBLAZING_ATK_STACK *
-    (1 * 1 / 10 + 2 * 1 / 10 + 3 * 1 / 10 + 4 * 1 / 10 + 5 * 1 / 10 + 6 * 1 / 10 + 7 * 1 / 10 + 8 * 3 / 10)
-
-  return {
-    display: () => (
-      <Flex vertical gap={10} >
-        <FormSwitch name='enemyProofOfDebtDebuff' text='Target proof of debt debuff' />
-        <FormSwitch name='numbyEnhancedState' text='Numby enhanced state' />
-        <FormSlider name='e1DebtorStacks' text='E1 debtor stacks' min={0} max={2} disabled={e < 1} />
-      </Flex>
-    ),
-    defaults: () => ({
-      enemyProofOfDebtDebuff: true,
-      numbyEnhancedState: true,
-      e1DebtorStacks: 2,
-    }),
-    precomputeEffects: (request) => {
-      const r = request.characterConditionals
-      const x = Object.assign({}, baseComputedStatsObject)
-
-      // Stats
-
-      // Scaling
-      x.BASIC_SCALING += basicScaling
-      x.SKILL_SCALING += skillScaling
-      x.SKILL_SCALING += (r.numbyEnhancedState) ? enhancedStateFuaScalingBoost : 0
-      x.FUA_SCALING += fuaScaling
-      x.FUA_SCALING += (r.numbyEnhancedState) ? enhancedStateFuaScalingBoost : 0
-
-      // Boost
-      x.FUA_VULNERABILITY += (r.enemyProofOfDebtDebuff) ? proofOfDebtFuaVulnerability : 0
-      x.ELEMENTAL_DMG += (request.enemyElementalWeak) ? 0.15 : 0
-      x.FUA_CD_BOOST += (e >= 1) ? 0.25 * r.e1DebtorStacks : 0
-
-      return x
-    },
-    calculateBaseMultis: (c, request) => {
-      const r = request.characterConditionals
-      const x = c.x
-
-      const hitMulti = (r.numbyEnhancedState) ? fuaEnhancedHitCountMulti : fuaHitCountMulti
-      const ashblazingFuaData = calculateAshblazingSet(c, request, hitMulti)
-      const ashblazingBasicData = calculateAshblazingSet(c, request, basicHitCountMulti)
-
-
-      x.BASIC_DMG += x.BASIC_SCALING * (x[Stats.ATK] - ashblazingBasicData.ashblazingAtk + ashblazingBasicData.ashblazingMulti)
-      x.FUA_DMG += x.FUA_SCALING * (x[Stats.ATK] - ashblazingFuaData.ashblazingAtk + ashblazingFuaData.ashblazingMulti)
-      x.SKILL_DMG = x.FUA_DMG
-
-      // Copy fua boosts to skill/basic
-      // BOOSTS get added, while vulnerability / def pen gets replaced (?)
-      x.SKILL_BOOST += x.FUA_BOOST
-      x.SKILL_CD_BOOST += x.FUA_CD_BOOST
-      x.SKILL_CR_BOOST += x.FUA_CR_BOOST
-      x.SKILL_VULNERABILITY = x.FUA_VULNERABILITY
-      x.SKILL_DEF_PEN = x.FUA_DEF_PEN
-      x.SKILL_RES_PEN = x.FUA_RES_PEN
-
-      x.BASIC_BOOST += x.FUA_BOOST
-      x.BASIC_CD_BOOST += x.FUA_CD_BOOST
-      x.BASIC_CR_BOOST += x.FUA_CR_BOOST
-      x.BASIC_VULNERABILITY = x.FUA_VULNERABILITY
-      x.BASIC_DEF_PEN = x.FUA_DEF_PEN
-      x.BASIC_RES_PEN = x.FUA_RES_PEN
-
-      // Her ult buff only applies to the skill/fua not basic
-      x.FUA_CD_BOOST += (r.numbyEnhancedState) ? enhancedStateFuaCdBoost : 0
-      x.SKILL_CD_BOOST += (r.numbyEnhancedState) ? enhancedStateFuaCdBoost : 0
-
-      // Her e6 only applies to skill/fua not basic
-      x.SKILL_RES_PEN += (e >= 6) ? 0.10 : 0
-      x.FUA_RES_PEN += (e >= 6) ? 0.10 : 0
-    }
-  }
-}
 
 function tingyun(e) {
   const skillAtkBoostMax = skill(e, 0.25, 0.27)
